feat: improve TLS handshakes and strategy scoring
- Correct TLS compression method length in FakePacketHelper
- Implement TLS noise injection for fragmented packets
- Update BypassConfig to use per-host strategy tracking
- Refactor test helpers for more realistic ClientHello generation

+++ app/src/main/java/com/aistudio/pinkproxy/fresh/FakePacketHelper.kt
@@ -439,7 +439,7 @@ object FakePacketHelper {
         )
         handDos.writeShort(ciphers.size); handDos.write(ciphers)
         // Compression Methods (1: null)
-        handDos.writeShort(1); handDos.writeByte(0)
+        handDos.writeByte(1); handDos.writeByte(0)
         
         // Extensions
         val extBaos = ByteArrayOutputStream(); val extDos = DataOutputStream(extBaos)
